Update patch set 1
Patch Set 1: Code-Review+2

Going to add the Android.mk in follow up CL?
Maybe a Readme.txt as outside of ChromeOS its not as clear what its for.

Also as this is a copy, we're gonna leave the history behind in the other repo, and leave the other repo as the upstream to this?

Patch-set: 1
Label: Code-Review=+2