Update patch set 2
Patch Set 2: Code-Review-1

There is a way to do that without going through zillions of gerrit CLs, we have done that when we merged platform/* into a single platform2 repo. It requires one time manual action from the gerrit instance admins (or grant you push access on these repos for a short period of time). avakulenko@ wiley@ and me are familiar with that process.

Having the history elsewhere is not as useful once the code moves a bit, files are renamed or refactored. You will end up checking the history in this repo, and in the other repo every time. Moving the history is not hard.

Squashing the history on truly "external" projects (third_party projects where the development happens somewhere else) is one thing... but doing that for Chromium projects where the OWNERS keep developing here is a different thing. Will this become the upstream minijail repo??

I'd like to agree on this process in the mailing list before we move forward with these changes. We will need to do the same for update_engine, shill, libchromeos, metrics, crash_reporter and cryptohome... some of which are also in "external/"

Patch-set: 2
Label: Code-Review=-1