minijail: Support entering an existing namespace.
When launching a full OS as the jailed process, it is useful to first be
able to configure a network namespace and start the new process in that
namespace.

This adds the "-e<net namespace file>" optional argument to -e.  It
allows, for example, passing "-e/var/run/netns/newns" to minijail0.

Also make the -v option work in a similar fashion, so there can
optionally be an existing namespace specified.  The user and pid
namespaces are dependent on each other so they will have the same option
added in a following commit.

Change-Id: I0613162072a1d14f10c58444c514f6d052c3d1e5
Signed-off-by: Dylan Reid <dgreid@chromium.org>