minijail: Add an option to drop caps without setting securebits
This option can be used if there is a desire to not set securebits but
still drop capabilities. The new -x option will be used to restrict
functionality in a container such as CAP_BLOCK_SUSPEND, without losing
other capabilities.

This also allows insecure caps to be dropped for a static binary because
the inheritance rules are less strict without the securebits set.

With '-x' root is still a privileged user, with '-c' root doesn't
default to having all privileges.  The -x option is not secure but it
works well when using capabilities to restrict functionality for policy
instead of security.

Change-Id: I1af8ebf916a807bed35e91ef5ddb3d7cf2264ecd
Signed-off-by: Dylan Reid <dgreid@chromium.org>