minijail: Allow static binaries in a bind mount to run
A previous commit placed a restriction on running static binaries that
are bind-mounted in to a chroot.  Remove that restriction by checking if
the binary path is in a bind mount and rebasing the path on to the bind
mount source path so that the executable can be accessed from outside
the chroot.

Change-Id: I801909df67c1bf18d48efcfd54c11aafe4c75e54
Signed-off-by: Dylan Reid <dgreid@google.com>