Update patch set 4
Patch Set 4:

The #ifdef still seems wrong, doesn't it? Now the log_syscalls struct is not gonna be defined on __ANDROID__ && __x86__?

I think you should invert the orders of the ifdefs, and have the arches first, and __ANDROID__ nested, that way until we add the syscalls for the other arches, the code still compiles.

Patch-set: 4