libminijail_generated: Use 32 bit syscalls in compat mode
On 64 bit architectures, 32 and 64 bit versions of libraries are
compiled to support compatibility with processes running in 32 bit
compatibility mode. Currently libminijail_generated.a is being created
with the 64 bit syscalls in both the 32 and 64 bit versions of the
library. Add target for TARGET_2ND_ARCH to use 32 bit syscalls in
the 32 bit libminijail_generated.a.

Change-Id: I34e22257b61a5c50deda2348e9cedc3304da0677