Pass clang CFLAGS into generate scripts
Add RS_TRIPLE_CFLAGS to ensure that we build with the proper defines on
targets like x86. This also changes all build targets to use the
proper 32/64-bit triples when creating their runtime libraries.

Without this change x86_64 system calls are incorrectly added
to libsyscalls.c on x86 builds. This passes in -D__i386__
for x86 and -D__x86_64__ for x86_64 targets.

Change-Id: I036565f65ce8a26554fe2a9f00a4d22834abea81