Update patch set 7
Patch Set 7: Code-Review-1

(1 comment)

I am removing my -2 so that I don't have to be disturbed again on vacation, but you can still count this as me -2'ing the direction of this patch currently. The steps need to be split out, so that the C/C++ compilation is done independent of the generated step. I think that Ying will be able to help explain anything else here.

Patch-set: 7
Label: Code-Review=-1