Update patch set 1
Patch Set 1:

> out of curiosity, why is this special cased for Android at all? The
 > same technique will work for all platforms, not just Android.

That's correct. The only reason to do it only for Android is that we have some seccomp filter policies on the Chrome OS side that restrict prctl() calls (from auto-generated syscall filters), and it's easier to unblock Android work by adding this and then modifying the policies, rather than the other way around.

Patch-set: 1