Make set_supplementary_gids abort on memory errors.
Consumers of this API usually cannot continue if this function fails,
since not adding supplementary groups would prevent the caller from
accessing resources. Simplify callers by aborting instead of returning
an error.

This would also prevent callers not checking the return value of the
function and not actually setting supplementary groups when they
expected to.

Bug: 26099611
Change-Id: Ib470e913d734ab4eac01b2aef3cdd4922d98e15a