Update patch set 4
Patch Set 4:

(2 comments)

I'm not sure this exactly is what you want, unless what you want is to avoid printing the warning message on old kernels.

This still leaves the bug in Minijail where we print the warning message but leave the seccomp_filter flag set. It seems like this would be a good opportunity to fix that.

I think the best way to do this is to introduce a new "int can_softfail" boolean variable in the function, that gets set in the following cases:

-When SECCOMP_SOFTFAIL is set OR
-When is_android() is true and the kernel version is < 3.8

That way you don't need to set the SECCOMP_SOFTFAIL flag enabled unconditionally on Android. 'Cause ideally, what we want here is to only enable the SECCOMP_SOFTFAIL flag in kernels < 3.8, but SECCOMP_SOFTFAIL being a compile-time flag, we don't really know what kernel we're gonna be running on by the time we set the flag. What do you think?

Patch-set: 4