Separate child process sync from user namespace
Syncing the child and parent was only done so that the uid/gid maps
could be setup.  Make this more general so that we can add a hook to do
further configuration before the child execs the jailed process.  A hook
will be added so that cgroups can be set up at this stage.

BUG=b/26549867
TEST=security_Minijail0

Change-Id: I81d512f351cfe459cd7af4c55263504d22b929fa
Signed-off-by: Dylan Reid <dgreid@chromium.org>