Add a post-fork hook function to libminjail
Add the ability to specify a callback function that will be called after
the child process is created, but before the jailed process is exec'd.
This will be used initially for adding the jailed process to cgroups,
ensuring that the jailed process and all children are in the desired
cgroups.

BUG=b/26549867
TEST=Use hook to set cgroup

Change-Id: I1c9ac54c5301d1f59cb650faa3a8e0e5a65b2f0b
Signed-off-by: Dylan Reid <dgreid@chromium.org>