add -T option to minijail0 for specifying ELF linkage type
This removes the need for minijail0 to access the program file before
trying to launch it, which allows it to be used for running programs
that are only accessible within a mount namespace specified with the -V
option.

It would still be preferable if -V worked correctly on its own (like
how -b, -k, and -C already do), but this is a less intrusive and
simpler immediate fix.

TEST=Use minijail0 -V -T to run program in pivot_root'd mount namespace

Bug: 26947503
Change-Id: I923ca87683b7fd8a60530946fad58018cfcd5125