defer checking ELF linkage type until after entering VFS namespace
This ensures that when we check the target program's ELF linkage, it
corresponds to the program we actually intend to execute.

Bonus: The check for libminijailpreload.so can now check that it's
accessible inside the VFS namespace too, and the access(filename,
X_OK) check tests the program's execute permissions according to our
final user credentials rather than the early credentials.

TEST=Use minijail0 -V to run a program in a pivot_root'd mount namespace

Bug: 26947503
Change-Id: Id9466731d63cf85283d8f38714831ce737e2b25d