Keep SHARED mount, if it is being introduced to a new namespace.
Currently, when a new mount namespace is created,
all existing mount points are forced to be re-marked
as PRIVATE.
To use SHARED/SLAVE mount point in mount namespaces,
this CL skips the marking, if the mount point is
intentionally being mounted as shared via commandline
flags.

Also, SHARED/SLAVE mounting does not work with
MS_BIND or MS_REMOUNT. Invoke mount(2) again in the
case.

Bug: 27304928
TEST=Ran minijail0 -v -k $SRC,$DEST,none,0x101000 which is \
     a bind mount with SHARED.

Change-Id: I3685815681a4d04f17d8a677a21fc6413f5e2154