Update patch set 2
Patch Set 2:

(4 comments)

Thank you for review.

> I'm a little bit uncomfortable with introducing a new behavior for -k flag without full understanding of who the users are.

It's open sourced so it is very difficult to track *all* users technically, unfortunately. Though, I think (almost) no users are affected, because;

The change of this CL is for the behavior with -k + MS_SHARED case only. Its current behavior is, 1) mark the mount point PRIVATE once, 2) bind mount in the new namespace with marking MS_SHARED. It means, it is marked only inside the new namespace. IMHO, it sounds useless...

Also, according to the log of the CL introducing -k option, the main target is only us at least at the moment. 

So, I think it is safer enough, IMHO.
Indeed, safer approach would be adding a new flag to control PRIVATE mount stuff, though it complicates minijail more. I'm ok with any way, if it satisfies our requirement, which is to use SHARED mount points in a new namespace. I'd defer the decision to Jorge.

Patch-set: 2
Label: Verified=0