Update patch set 2
Patch Set 2:

(1 comment)

The whole point of mount namespaces is to keep mounts in one namespace separate from mounts in other namespaces. I'm not sure if doing this in Minijail is the right way to go. Why does the sdcard container need to be in a mount namespace if it's implementing a mountpoint that needs to be shared?

Yes, this functionality was first implemented for Lakitu -- but the reality is that before that, we had a bug. Mounts inside a mount namespace should never leak. Minijail is used on other OSs that are not Chrome OS, and we cannot depend on things being mounted MS_PRIVATE beforehand.

Patch-set: 2
Label: Verified=0