Update patch set 1
Patch Set 1: Code-Review+2

so we're on the same page, this just makes the code work on pre-4.3 kernels (which added PR_CAP_AMBIENT_RAISE), but it also means that 4.3+ kernels won't have that cap locked down.

looking at the history, this code requires linux 2.6.26+, but we don't care about stuff that old.  maybe we should put a note in the source/readme that states the min required version (or maybe throws an #error).

Patch-set: 1
Reviewer: Gerrit User 1040868 <1040868@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2