Update patch set 1
Patch Set 1:

when the next version adds another perm, are we just going to have a perpetual march of trying different bitmasks?

the caps code exports info via /proc so we can get this at runtime. is there a similar knob for this? or maybe we can look at extending the API in some way to support what are need?

Patch-set: 1