Allow process to enter existing user namespace
This adds a new option to minijail to enter an
existing user namespace. This allows for sharing
of namespaces between processes in situations
where setns requires CAP_SYS_ADMIN capabilities
in the user namespace that created the namespace
to be shared.

BUG=chromium:629685
TEST=Run a process with minijail in new user and
net namespaces. Run a new process that uses the
same user and net namespaces as the first process.
Set up a loopback interface in the first process
and verify that the same interface is visible in
the second process.

Change-Id: I75d81864f5b7ef7f4ca332917615a4733f8a5516