Add the possibility to extend minijail with callbacks
Some functionality is very context specific and as such should
be left to the calling application to handle in a manner that is
appropriate. What is required from minijail is to ensure that
the functionality is called at the correct phase of the containment.

Callback functions will allow minijail to be more extensible. This
initial version, adds callback registering and invoking
functionality for selinux.

Change-Id: Ia12dcca7b92324f01ef5c938095f9b25dbe5c3f8
Signed-off-by: Brian McGillion <brian.mcgillion@darkmatter.ae>