Update patch set 1
Patch Set 1:

FTR I noticed that job control doesn't work when I minijail a chrooted bash from a shell.  The reason for this is isatty(3) fails, most likely for /dev/fd/{1,2,3} which are symlinks to /dev/pts/%d which is outside the chroot but not inside.  Is there an easy way to put these in as well?  The naive method of inspecting /proc/self/fd pre-jail and mknod these things post-jail comes to mind.

Patch-set: 1