Update patch set 4
Patch Set 4:

If a buggy snprintf(3) decides to return a negative number now, we'll treat it as a large positive number and therefore a massive error (as it should be).  This is because of the signed to unsigned conversion is defined as "plus U*_MAX + 1" to get the least unsigned integer congruent to the source (mod 2^n).

Patch-set: 4
Label: Verified=0