Update patch set 5
Patch Set 5:

(2 comments)

this seems like a lot of work to simply parse a number, especially to try and make sure we get a size_t one (when no valid number would actually need that size).  plus, this only applies to the tmpfs spec when this could show up in more places.

counter proposal: update util.c's parse_single_constant to handle suffixes, or add a new helper to util.c that uses parse_single_constant for the numeric part and follows it up with a suffix check.

Patch-set: 5