Add a flag to drop access to the session keyring
Now that chrome os is moving over to ext4 based directory encryption, the
encryption keys are stored in the session keyring.  Applications that don't need
to access the encrypted user data directory don't need access to this keyring.
Add a new flag to create a new anonymous session keyring for those applications.

Bug: crbug.com/682419
TEST=autotest in a later CL

Change-Id: I3cb8f120d9f4891d9a13f7fe342b0388e9975605
Signed-off-by: Chirantan Ekbote <chirantan@google.com>