Update patch set 1
Patch Set 1:

This will break Minijail on Android. You would have to modify Android.mk as well, and make sure keyutils is available on Android.

Thinking about this a little bit more, does it make sense to add all of the keyutils library just for a single keyctl_join_session_keyring(NULL) call? Do we know if the kernel interface is easy to use?

Patch-set: 1