Update patch set 1
Patch Set 1:

> > Patch Set 1:
 > >
 > > This will break Minijail on Android. You would have to modify
 > Android.mk as well, and make sure keyutils is available on Android.
 > >
 > > Thinking about this a little bit more, does it make sense to add
 > all of the keyutils library just for a single keyctl_join_session_keyring(NULL)
 > call? Do we know if the kernel interface is easy to use?
 > 
 > It's pretty straightforward.  I can just make the syscall directly
 > if we're ok with defining the command number locally in the file.

I think that would make sense.

And if we want to make this more resilient, we can always depend on the library in a test, and have the test check that the value is not changing (for example, by compiling a small program with a static_assert).

Patch-set: 1