Update patch set 1
Patch Set 1:

> > Patch Set 1:
 > >
 > > > Patch Set 1:
 > > >
 > > > This does fail on Android if you run the 32-bit unit test (I
 > just tested locally), but unfortunately the devices in the lab are
 > all 64-bit devices =(
 > >
 > > Are you sure you have this CL in your tree?  Because I tested on
 > x86-alex before uploading and it was broken before this change but
 > was fixed with this change applied.
 > 
 > Even looking at the code, the return value _should_ be -ERANGE when
 > __WORDSIZE==32 and -EINVAL when __WORDSIZE==64.

Not sure what you mean. Your complaint was that the Android presubmit system had not caught the fact that the new unit test was not passing on 32-bit systems. I was confirming the fact that the test *was* failing on a 32-bit Android device, but that the presubmit system did not have these devices. I was also confirming that the fix was working on 32-bit Android devices. What does that have to do with having or not having the CL on my tree?

Patch-set: 1