Add option '-z' to pass all signals to child
When minijail0 is called with the `-z` option a signal handler for
every signal will be setup in the minijail0 process that simply sends
the signal along to the spawned child.

This is used to make daemon management (e.g. with `sv`) easier, since
otherwise SIGTERM on minijail0 just orphans the true service.

Change-Id: Ide0471511a4242839e3c8a34795e5a297b6ef57a