Update patch set 5
Patch Set 5:

when minijail is acting as the init inside of the pid namespace, it needs to change its behavior depending on whether the signal is coming from outside the pid namespace or from inside.  if it's outside, it needs to kill its children and then kill itself.  if it's inside, it needs to ignore the signal.

i don't think we need to implement that right now, but we need to keep it in mind so we can share the logic.

one other special case: you prob want to set SIGCHLD to SIG_DFL.  if minijail is monitoring a child pid, it's doing so via wait.

Patch-set: 5
Reviewer: Gerrit User 1040868 <1040868@85c56323-6fa9-3386-8a01-6480fb634889>