Update patch set 1
Patch Set 1:

Sorry that the context is missing.

In Treble, everything that is in /vendor partition are built against the subset of platform libraries that we call VNDK. libminijail is used by some libraries in vendor and thus need to be in VNDK. When a lib is in VNDK, two versions are built. One for platform itself (installed to /system), the other for vendor (installed to /vendor).

Please refer to this doc as well: go/build-system-vendor-split

Patch-set: 1