Update patch set 3
Patch Set 3:

Is the list needed?

Is a callback even needed or should we add set_selinux_context instead?  

The callback is limited in usefulness because it runs after much of the jailing has already happened, it requires all the functions to be allowed by the seccomp policy for example.

Patch-set: 3
CC: Gerrit User 1067948 <1067948@85c56323-6fa9-3386-8a01-6480fb634889>