minijail: Add a pre-start hook in the parent process
This change adds a hook in the parent process that is run before the
child calls execve(2). The signature of the hooks was also changed to
add the pid of the child, since the parent has no other way of knowing
the child pid by the time the hook is invoked.

Bug: 63904978
Test: make tests
Change-Id: I83a901dd6fe03bd229fed8786b6d38b115ab3704