abort when bind mounting a non-existent source path
If you try to pass a bogus path like -b /asdf,/asdf, minijail doesn't
mind and goes ahead and creates the destination (which also creates
the source), and then does a bind mount.  We should instead abort --
if you really want to bind mount a new directory or file, the daemon
should explicitly create the path correctly.

The current behavior is also a little idiosyncratic: if the source
and dest are the same, there's no error, but if you try to mount to
a different path (-b /asdf,/foo), it'll fail.  Or if you try to use
a chroot/pivot root, it'll fail.

Bug: None
Test: None

Change-Id: I26310ba45b8e463533485de879a19e578d66b0e6