Update patch set 2
Patch Set 2:

after some discussion in another CrOS CL, the desire to use /dev with a chroot came up, but the previous code didn't work with that at all.  it was easy to fix that (just change the /dev to be under the chroot path), but that wouldn't help too much more -- we'd silently fail to properly bind mount /dev paths like /dev/log.

so i had to rearchitect this significantly.  i don't think the additional complexity is a huge deal considering the benefits.  ptal!

Patch-set: 2
Reviewer: Gerrit User 1040868 <1040868@85c56323-6fa9-3386-8a01-6480fb634889>