Update patch set 1
Patch Set 1:

We were re-implementing a bunch of stuff provided by minijail run in crosvm because we wanted to put a child process in a usernamespace but not exec a new program.

Does this seem like a reasonable way to enable that situation to use minijail_run functionality?

Thanks,

Dylan

Patch-set: 1
Reviewer: Gerrit User 1067948 <1067948@85c56323-6fa9-3386-8a01-6480fb634889>