Update patch set 1
Patch Set 1:

this is always a waste of resources.  this is returning from main which means the next step is to exit which means the kernel will take care of releasing all our resources for us.

if you want to avoid leak-at-exit warnings from asan, put this logic behind the __SANITIZE_ADDRESS__ define.

Patch-set: 1
CC: Gerrit User 1040868 <1040868@85c56323-6fa9-3386-8a01-6480fb634889>