Add a way to specify the uid/gid in userns
This change adds a way to specify the uid/gid that will be used when
entering the user namespace. This is used to fix an issue where calling
open(2) on a file on the tmpfs created by minijail would return
EOVERFLOW[1]. An easy way to reproduce is running this on a 4.8 kernel
(or Ubuntu Xenial, which has this change backported):

  $ ./minijail0 -T static -Ut -- /bin/bash -c 'touch /tmp/foo'

This change allows the caller to explicitly choose the arguments to
setsuid(2)/setsgid(2) when entering the userns, to avoid the above
issue.

1: More information about the bug here:
https://bugs.launchpad.net/ubuntu/+source/linux/+bug/1659087

Bug: None
Test: make tests
Change-Id: I393daaf8c2b2355e33c75a908345bb03f1980271