Mark libminijail as VNDK
This commit marks libminijail as a VNDK library so that the vendor
variant can be installed into /system/lib[64]/vndk instead of
/vendor/lib[64].

This library was not a VNDK library in Android O because its Android.mk
hasn't been converted to Android.bp at the time.  Thus, we were asking
vendors to use libminijail_vendor instead and omx implementation has to
use libavservice_minijail_vendor.  Since Android.mk has been converted
to Android.bp, it would be great to mark this library as VNDK library so
that libminijail can be updated through system-only OTAs.

Bug: 63866913
Test: Build sailfish and libminijail.so should be installed into
/system/lib[64]/vndk/libminijail.so instead of
/vendor/lib[64]/libminijail.so.

Change-Id: Ia03d2ec7ae43f3a316815087fd192e0b8ab0456f