Update patch set 1
Patch Set 1:

i'm not quite following the logic here.  there was a problem or something in the past, but that's been resolved now, so why do we still care ?  libminijail needs to be kept in sync with the minijail program, so updating one and not the other would be bad, and they both should be system binaries rather than something vendors get to mess with.

Patch-set: 1
Reviewer: Gerrit User 1040868 <1040868@85c56323-6fa9-3386-8a01-6480fb634889>