Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> i'm not quite following the logic here.  there was a problem or something in the past, but that's been resolved now, so why do we still care ?  libminijail needs to be kept in sync with the minijail program, so updating one and not the other would be bad, and they both should be system binaries rather than something vendors get to mess with.

I am not sure which minijail program you are referring.  If you are referring the `minijail0` executable specified in the Android.bp file, then it seems not being installed on the device.  If you are referring the users of libminijail, then there is one known case in the vendor partition.

Currently, android.hardware.media.omx@1.0-service (mediacodec) is installed in the vendor partition.  It is depending on /vendor/lib[64]/libavservices_minijail_vendor.so and indirectly depending on /vendor/lib[64]/libminijail.so (or previously /vendor/lib[64]/libminijail_vendor.so).

The motivation of this commit is to move /vendor/lib[64]/libminijail.so back to /system/lib[64]/vndk/libminijail.so.  There will be a snapshot for shared libraries in /system/lib[64]/vndk, thus it should be ABI-compatible.  The updates will only apply to the snapshot if there are some security patches and such security patches can be applied without changing ABIs.

p.s. If security updates break the ABIs, then both system and vendor partition should be updated (this is similar to the world we have today).

Patch-set: 1
Reviewer: Gerrit User 1014807 <1014807@85c56323-6fa9-3386-8a01-6480fb634889>