Update patch set 1
Patch Set 1:

the only reason you want libminijail in vndk is because, at some point in the past, some vendors built & installed a copy of libminijail_vendor.so in /vendor/ and you want to now update them ?

but for devices that start on Android P, this wouldn't come up ?

what i'm getting at is that it'd be nice if we could figure out how to keep minijail as a system package and perhaps deploy the vendor hack only for builds that actually need it, instead of making all systems from now on need this as a vendor bit.  that way we eventually converge on the sane resting state of nothing vendor related.

Patch-set: 1
Reviewer: Gerrit User 1040868 <1040868@85c56323-6fa9-3386-8a01-6480fb634889>