create parent paths of target mounts as needed
Currently if you want to bind mount a single subdir, you have to make
sure to create the full parent directory chain.  For example, if you
want /var/lib/timezone/ but not the rest of /var, you have to do:
	-k none,/var,tmpfs
	-k none,/var/lib,tmpfs
	-b /var/lib/timezone/
For every additional subdir, you need to add another -k option just to
do an [effective] mkdir with a tmpfs mount.

The current -k/-b behavior is to run mkdir if the target doesn't already
exist, but only for the final target.  Lets extend it to also create any
missing parent paths, so now only the base path needs to be writable:
	-k none,/var,tmpfs
	-b /var/lib/timezone/

Bug: None
Test: `minijail0 --profile minimalistic-mountns -k none,/var,tmpfs -b /var/lib/timezone /bin/date` works
Change-Id: I7f36bcb445ce40ed66a9403a4ee1c1fe3f9e5ea8