Update patch set 1
Patch Set 1:

i don't have a build log handy, but it was when building the code to run on the host.  AOSP's host prebuilt is ancient: prebuilts/gcc/linux-x86/host/x86_64-linux-glibc2.15-4.8/.  the kernel headers used to build that are from 3.2.14 (look at the linux/version.h in that tree).

it's not related to the APCT shuffling as those are only test-time failures and i'm seeing build-time failures.

Patch-set: 1
Reviewer: Gerrit User 1040868 <1040868@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0