Use uintptr_t for the arguments in minijail_rlimit()
This changes the type used to pass the value of the limit from uint32_t
to uintptr_t (since rlim_t is typedef'd to __syscall_ulong_t, which is
the type used to pass arguments to syscalls, which should the same size
of a pointer). It also provides a bit of API backwards-compatibility in
case anybody tried to cast a value (such as RLIM_INFINITY) to a
uint32_t.

Bug: None
Test: minijail_rlimit(j, RLIMIT_STACK, RLIM_INFINITY, RLIM_INFINITY);
Change-Id: I6bbf956974c7f28a2e0bb8ad3efb4d25c5675593