minijail: Move the call to abort(2) into do_fatal_log()
This change moves the invocation to abort(2) into do_fatal_log(). This
causes the compiler to avoid simply inserting a branch to a common place
in large functions (such as minijail_enter()) that eventually calls
abort(2), which in turn should make the breakpad-based stack unwinder be
able to more accurately pinpoint the failing line instead of just
blaming the whole function.

Bug: 111271997
Test: objdump -d minijail0 | grep abort@plt | wc -l  # 1
Test: objdump -d libminijailpreload.so | grep abort@plt | wc -l  # 2

Change-Id: I770c8f1a06736100506053685f3e380a2576a80b